Implemented bread mode
